Wrap the whole project to be a package.
As we are growing rapidly, pooling all files in the root folder is no
longer a good idea, and we need to separate non-infra files, such as
alphas and crawlers, from infra as well. To handle complicated
cross-import issues following the re-organization, we must convert our
project from a set of scripts to a package.

In this commit, org-level changes include:

  1. All crawelers (excluding the template, crawler.py) are moved to crawlers/
  2. All data will be saved in data/ (path registered in constants.py)
  3. All import should be absolute path based on the project root folder.

Besides, path generated by the old pylint.yml will include only file
names but not full relative path. This will lose package context and
lead to pylint errors, so its command is modified to fix the bug.