Adjust alpha_formula to fit initial condition
The argument, date, of __init__ in AlphaFormula is the simulation start
date, which is also what we pass to "generate" in the first round.
Thus, in __init__ we should set initial values to the state of the day
BEFORE start_date; otherwise, new available data on start_date (data[-1])
will be calculated twice.